requirements.txt
absl-py==0.9.0
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
beautifulsoup4==4.6.0
blessings==1.7
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
devscripts===2.17.12ubuntu1.1
distro-info===0.18ubuntu0.18.04.1
gast==0.3.3
google-pasta==0.2.0
gpg==1.10.0
greenlet==0.4.12
grpcio==1.27.2
h5py==2.10.0
html5lib==0.999999999
httpie==0.9.8
httplib2==0.9.2
idna==2.6
joblib==0.14.1
kazam==1.4.5
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
mach==0.6
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.0
matplotlib==3.2.0
meson==0.51.0
mozfile==2.0.0
mozinfo==1.1.0
mozprocess==1.0.0
msgpack==0.5.6
neovim==0.2.0
netifaces==0.10.4
numpy==1.18.1
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.11.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.6
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-ldap==3.2.0
python-magic==0.4.16
pytz==2018.3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==5.3
remarkable==1.87
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.22.2.post1
scipy==1.4.1
scour==0.36
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.5
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.4
urllib3==1.22
usb-creator==0.3.3
vboxapi==1.0
wadllib==1.3.2
webencodings==0.5
Werkzeug==1.0.0
wrapt==1.12.1
xkit==0.0.0
youtube-dl==2018.3.14
zope.interface==4.3.2
